+++ src/widget/widget.rs
@@ -66,6 +66,47 @@ pub enum WidgetConfig {
     BorderWidth { thickness: u8 },
 }
 
+/// This structure is used for the configuration store of `Widget` settings.  It contains its
+/// own structure internally, so all that is used inside extended `Widget` objects is a simple
+/// instantiation of a new `Configurable` object as part of your extension.
+pub struct Configurable {
+    config: HashMap<ConfigKey, WidgetConfig>,
+}
+
+/// Implementation of the `Configurable` object.  Contains methods to extend the `HashMap` that
+/// is used for underlying storage.
+impl Configurable {
+    /// Creates a new instance of this object.
+    pub fn new() -> Self {
+        Self {
+            config: HashMap::new(),
+        }
+    }
+
+    /// Sets a configuration key by its `ConfigKey` ID, assigning a new `WidgetConfig` value
+    /// to that key.
+    pub fn set(&mut self, key: ConfigKey, value: WidgetConfig) {
+        self.config.insert(key, value);
+    }
+
+    /// Retrieves an `Option<&WidgetConfig>` for the key specified.  If the key does not exist,
+    /// a `None` is returned.
+    pub fn get(&self, key: ConfigKey) -> Option<&WidgetConfig> {
+        self.config.get(&key)
+    }
+
+    /// Removes the value for the specified key, if one exists.
+    pub fn remove(&mut self, key: ConfigKey) {
+        self.config.remove(&key);
+    }
+
+    /// Indicates whether or not a `Configurable` store contains a value for the specified key.
+    /// Returns `true` if one is stored, `false` otherwise.
+    pub fn contains_key(&self, key: ConfigKey) -> bool {
+        self.config.contains_key(&key)
+    }
+}
+
 /// Implementable trait that is used by every `Widget`.  These are the public methods,
 /// and a function _may_ override them.
 ///
@@ -91,16 +132,17 @@ pub trait Widget {
     /// # use pushrod::widget::widget::Widget;
     /// # use pushrod::widget::widget::WidgetConfig;
     /// # use pushrod::widget::widget::ConfigKey;
+    /// # use pushrod::widget::widget::Configurable;
     /// # use std::collections::HashMap;
     /// # use std::cell::RefCell;
     /// struct MyWidget {
-    ///   config: RefCell<HashMap<ConfigKey, WidgetConfig>>
+    ///   config: Configurable,
     /// }
     ///
     /// impl MyWidget {
     ///   fn new() -> Self {
     ///     Self {
-    ///       config: RefCell::new(HashMap::new()),
+    ///       config: Configurable::new(),
     ///     }
     ///   }
     /// }
@@ -111,17 +153,18 @@ pub trait Widget {
     /// ```
     /// # use pushrod::widget::widget::Widget;
     /// # use pushrod::widget::widget::WidgetConfig;
+    /// # use pushrod::widget::widget::Configurable;
     /// # use std::collections::HashMap;
     /// # use std::cell::RefCell;
     /// # use pushrod::core::point::Point;
     /// struct MyWidget {
-    ///   config: RefCell<HashMap<u8, WidgetConfig>>
+    ///   config: Configurable,
     /// }
     ///
     /// impl Widget for MyWidget {
     ///
-    ///   fn get_config(&mut self) -> &RefCell<HashMap<u8, WidgetConfig>> {
-    ///     &self.config
+    ///   fn get_config(&mut self) -> &mut Configurable {
+    ///     &mut self.config
     ///   }
     ///
     ///  fn mouse_entered(&mut self, widget_id: i32) {}
@@ -132,11 +175,11 @@ pub trait Widget {
     ///
     /// This uses a `RefCell`, since configurations require a mutable reference to the HashMap
     /// that stores the configs.
-    fn get_config(&mut self) -> &RefCell<HashMap<ConfigKey, WidgetConfig>>;
+    fn get_config(&mut self) -> &mut Configurable;
 
     /// Sets a configuration object by its key.
     fn set_config(&mut self, key: u8, value: WidgetConfig) {
-        self.get_config().borrow_mut().insert(key, value);
+        self.get_config().set(key, value);
     }
 
     /// Indicates that a widget needs to be redrawn/refreshed.
@@ -146,40 +189,50 @@ pub trait Widget {
 
     /// Clears the invalidation flag.
     fn clear_invalidate(&mut self) {
-        self.get_config().borrow_mut().remove(&CONFIG_INVALIDATE);
+        self.get_config().remove(CONFIG_INVALIDATE);
     }
 
     /// Checks to see whether or not the widget needs to be redrawn/refreshed.
     fn is_invalidated(&mut self) -> bool {
-        self.get_config().borrow().contains_key(&CONFIG_INVALIDATE)
+        self.get_config().contains_key(CONFIG_INVALIDATE)
     }
 
-    /// Sets the `Point` of origin for this widget.  Invalidates the widget afterward.
-    fn set_origin(&mut self, point: Point) {
-        self.set_config(CONFIG_ORIGIN, WidgetConfig::Origin { point });
+    /// Sets the `Point` of origin for this widget, given the X and Y origin points.  Invalidates the widget afterward.
+    fn set_origin(&mut self, x: i32, y: i32) {
+        self.set_config(
+            CONFIG_ORIGIN,
+            WidgetConfig::Origin {
+                point: Point { x, y },
+            },
+        );
         self.invalidate();
     }
 
     /// Retrieves the `Point` of origin for this object.
     /// Defaults to origin (0, 0) if not set.
     fn get_origin(&mut self) -> Point {
-        match self.get_config().borrow().get(&CONFIG_ORIGIN) {
+        match self.get_config().get(CONFIG_ORIGIN) {
             Some(WidgetConfig::Origin { ref point }) => point.clone(),
             None => make_origin_point(),
             _ => make_origin_point(),
         }
     }
 
-    /// Sets the `Size` for this widget.  Invalidates the widget afterward.
-    fn set_size(&mut self, size: crate::core::point::Size) {
-        self.set_config(CONFIG_SIZE, WidgetConfig::Size { size });
+    /// Sets the `Size` for this widget, given a width and height.  Invalidates the widget afterward.
+    fn set_size(&mut self, w: i32, h: i32) {
+        self.set_config(
+            CONFIG_SIZE,
+            WidgetConfig::Size {
+                size: crate::core::point::Size { w, h },
+            },
+        );
         self.invalidate();
     }
 
     /// Retrieves the `Size` bounds for this widget.
     /// Defaults to size (0, 0) if not set.
     fn get_size(&mut self) -> crate::core::point::Size {
-        match self.get_config().borrow().get(&CONFIG_SIZE) {
+        match self.get_config().get(CONFIG_SIZE) {
             Some(WidgetConfig::Size { ref size }) => size.clone(),
             None => make_unsized(),
             _ => make_unsized(),
@@ -195,9 +248,10 @@ pub trait Widget {
     /// Retrieves the color of this widget.
     /// Defaults to white color `[1.0; 4]` if not set.
     fn get_color(&mut self) -> types::Color {
-        if self.get_config().borrow().contains_key(&CONFIG_COLOR) {
-            match self.get_config().borrow()[&CONFIG_COLOR] {
-                WidgetConfig::Color { color } => color,
+        if self.get_config().contains_key(CONFIG_COLOR) {
+            match self.get_config().get(CONFIG_COLOR) {
+                Some(WidgetConfig::Color { ref color }) => [color[0], color[1], color[2], color[3]],
+                None => [1.0; 4],
                 _ => [1.0; 4],
             }
         } else {
@@ -246,7 +300,7 @@ pub trait Widget {
 /// This is the `BaseWidget`, which contains a top-level widget for display.  It does
 /// not contain any special logic other than being a base for a display layer.
 pub struct BaseWidget {
-    config: RefCell<HashMap<ConfigKey, WidgetConfig>>,
+    config: Configurable,
 }
 
 /// Implementation of the constructor for the `PushrodBaseWidget`.  Creates a new base widget
@@ -254,7 +308,7 @@ pub struct BaseWidget {
 impl BaseWidget {
     pub fn new() -> Self {
         Self {
-            config: RefCell::new(HashMap::new()),
+            config: Configurable::new(),
         }
     }
 }
@@ -280,8 +334,8 @@ impl BaseWidget {
 /// #
 ///    let mut base_widget = BaseWidget::new();
 ///
-///    base_widget.set_origin(Point { x: 100, y: 100 });
-///    base_widget.set_size(pushrod::core::point::Size { w: 200, h: 200 });
+///    base_widget.set_origin(100, 100);
+///    base_widget.set_size(200, 200);
 ///    base_widget.set_color([0.5, 0.5, 0.5, 1.0]);
 ///
 ///    // Widgets must be boxed, as they are trait objects.
@@ -291,17 +345,17 @@ impl BaseWidget {
 ///
 ///    let mut base_widget_2 = BaseWidget::new();
 ///
-///    base_widget_2.set_origin(Point { x: 125, y: 125 });
-///    base_widget_2.set_size(pushrod::core::poiNt::Size { w: 100, h: 100 });
+///    base_widget_2.set_origin(125, 125);
+///    base_widget_2.set_size(100, 100);
 ///    base_widget_2.set_color([0.75, 0.75, 0.75, 1.0]);
 ///
 ///    // Add the second widget to the top level base widget.
-///    let widget_id_2 = pushrod_window.add_widget_to_parent(Box::new(base_widget_2, widget_id);
+///    let widget_id_2 = pushrod_window.add_widget_to_parent(Box::new(base_widget_2), widget_id);
 /// # }
 /// ```
 impl Widget for BaseWidget {
-    fn get_config(&mut self) -> &RefCell<HashMap<ConfigKey, WidgetConfig>> {
-        &self.config
+    fn get_config(&mut self) -> &mut Configurable {
+        &mut self.config
     }
 
     fn mouse_entered(&mut self, widget_id: i32) {
@@ -319,4 +373,3 @@ impl Widget for BaseWidget {
         );
     }
 }
-